# backend/app/web/static/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "股票FinCore AI API",
    "description": "FinCore AI Stock Analysis System REST API",
    "version": "2.1.0"
  },
  "host": "localhost:8888",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "securityDefinitions": {
    "APIKeyHeader": {
      "type": "apiKey",
      "name": "X-API-KEY",
      "in": "header"
    },
    "HMACSignature": {
      "type": "apiKey",
      "name": "X-HMAC-SIGNATURE",
      "in": "header"
    }
  },
  "paths": {
...
"/api/analysis_status/{task_id}": {
      "get": {
        "summary": "Get analysis task status",
        "description": "Retrieves the status and result of an asynchronous analysis task.",
        "produces": [
          "application/json"
        ],
        "security": [
          {
            "APIKeyHeader": [],
            "HMACSignature": []
          }
        ],
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Task status",
            "schema": {
              "$ref": "#/definitions/TaskStatusResponse"
            }
          },
          "404": {
            "description": "Task not found",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      }
    },
...
    "/api/fundamental_analysis": {
      "post": {
        "summary": "基本面分析",
        "description": "获取股票的基本面分析结果",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "stock_code": {
                  "type": "string",
                  "example": "600519"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功获取基本面分析结果"
          }
        }
      }
    },
    "/api/capital_flow": {
      "post": {
        "summary": "资金流向分析",
        "description": "获取股票的资金流向分析结果",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "stock_code": {
                  "type": "string",
                  "example": "600519"
                },
                "days": {
                  "type": "integer",
                  "example": 10
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功获取资金流向分析结果"
          }
        }
      }
    },
    "/api/scenario_predict": {
      "post": {
        "summary": "情景预测",
        "description": "获取股票的多情景预测结果",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "stock_code": {
                  "type": "string",
                  "example": "600519"
                },
                "market_type": {
                  "type": "string",
                  "example": "A"
                },
                "days": {
                  "type": "integer",
                  "example": 60
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功获取情景预测结果"
          }
        }
      }
    },
...
    "/api/risk_analysis": {
      "post": {
        "summary": "风险分析",
        "description": "获取股票的风险分析结果",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "stock_code": {
                  "type": "string",
                  "example": "600519"
                },
                "market_type": {
                  "type": "string",
                  "example": "A"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功获取风险分析结果"
          }
        }
      }
    },
    "/api/portfolio_risk": {
      "post": {
        "summary": "投资组合风险分析",
        "description": "获取投资组合的整体风险分析结果",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "portfolio": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "stock_code": {
                        "type": "string"
                      },
                      "weight": {
                        "type": "number"
                      },
                      "market_type": {
                        "type": "string"
                      }
                    }
                  },
                  "example": [
                    {
                      "stock_code": "600519",
                      "weight": 30,
                      "market_type": "A"
                    },
                    {
                      "stock_code": "000858",
                      "weight": 20,
                      "market_type": "A"
                    }
                  ]
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功获取投资组合风险分析结果"
          }
        }
      }
    },
    "/api/index_analysis": {
      "get": {
        "summary": "指数分析",
        "description": "获取指数的整体分析结果",
        "parameters": [
          {
            "name": "index_code",
            "in": "query",
            "required": true,
            "type": "string",
            "example": "000300"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "type": "integer",
            "example": 30
          }
        ],
        "responses": {
          "200": {
            "description": "成功获取指数分析结果"
          }
        }
      }
    },
    "/api/industry_analysis": {
      "get": {
        "summary": "行业分析",
        "description": "获取行业的整体分析结果",
        "parameters": [
          {
            "name": "industry",
            "in": "query",
            "required": true,
            "type": "string",
            "example": "银行"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "type": "integer",
            "example": 30
          }
        ],
        "responses": {
          "200": {
            "description": "成功获取行业分析结果"
          }
        }
      }
    },
    "/api/industry_compare": {
      "get": {
        "summary": "行业比较",
        "description": "比较不同行业的表现",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "type": "integer",
            "example": 10
          }
        ],
        "responses": {
          "200": {
            "description": "成功获取行业比较结果"
          }
        }
      }
    }
  },
  "definitions": {
...
    "TaskStatusResponse": {
        "type": "object",
        "properties": {
            "task_id": {
                "type": "string"
            },
            "status": {
                "type": "string"
            },
            "result": {
                "type": "object"
            }
        }
    },
...
    "Error": {
        "type": "object",
        "properties": {
            "error": {
                "type": "string"
            }
        }
    },
...
  }
}